Extract translations into separate files
This way if an application does not need a particular translation, the
file can just be omitted and the overall file size will be reduced.

Adding new translations becomes easier this way, too.

Many modern web frameworks support asset file bundling (concatenation
and minification) and multiple files should not be an issue.